refactor: 🎨 rename mod
